Migrate unit tests from JUnit 4 to 5
This leverages the bill of materials to simplify JUnit dependency
management. It also removes all transitive dependencies on JUnit 4.
Finally, it updates all tests to use JUnit 5.

Note that for the test
`XSyncTest::xSyncWithDifferentMutexFactoriesDoesntLock`, it was
expecting a `Throwable` to be thrown at any time. In practice,
sometimes, this was an NPE and sometimes it was a failed assertion. I
fixed the NPE, negated the assertion, and removed the expected
exception.

Resolves: #7