Fix expected inconsistency on CI build
The `XSyncTest::xSyncWithDifferentMutexFactoriesDoesntLock` exhibited
different behaviour from both my IDE run and command line run. There was
no contention on the CI build. I modified the operation to increase the
likelihood of inconsistent reads. If it still does not work, I can
insert a `Thread.yield()` or worst case, a short sleep between the read
and the write.

Resolves: #7